tools, vm: new option to specify kpageflags file
page-types currently hardcodes /proc/kpageflags as the file to parse. 
This works when using the tool to examine the state of pageflags on the
same system, but does not allow storing a snapshot of pageflags at a given
time to debug issues nor on a different system.

This allows the user to specify a saved version of kpageflags with a new
page-types -F option.

Link: http://lkml.kernel.org/r/alpine.DEB.2.10.1801301458180.153857@chino.kir.corp.google.com
Signed-off-by: David Rientjes <rientjes@google.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Konstantin Khlebnikov <koct9i@gmail.com>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>