sysctl: warn when a clamped sysctl parameter is set out of range
Even with clamped sysctl parameters, it is still not that straightforward
to figure out the exact range of those parameters.  One may try to write
extreme parameter values to see if they get clamped.  To make it easier, a
warning with the expected range will now be printed in the kernel ring
buffer when a clamped sysctl parameter receives an out of range value.

Link: http://lkml.kernel.org/r/1519059231-2456-3-git-send-email-longman@redhat.com
Signed-off-by: Waiman Long <longman@redhat.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Cc: Manfred Spraul <manfred@colorfullife.com>
Cc: "Luis R. Rodriguez" <mcgrof@kernel.org>
Cc: Kees Cook <keescook@chromium.org>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>