kernel/sysctl.c: add flags to support min/max range clamping
When minimum/maximum values are specified for a sysctl parameter in the
ctl_table structure with proc_dointvec_minmax() handler, updates to that
parameter will fail with error if the given value is outside of the
required range.

There are use cases where it may be better to clamp the value of the
sysctl parameter to the given range without failing the update, especially
if the users are not aware of the actual range limits.  Reading the value
back after the update will now be good practice to see if the provided
value exceeds the range limits.

To provide this less restrictive form of range checking a new flags field
is added to the ctl_table structure.  The new field is a 16-bit value that
just fits into the hole left by the 16-bit umode_t field without
increasing the size of the structure.

When the CTL_FLAGS_CLAMP_RANGE flag is set in the ctl_table entry, any
update from userspace will be clamped to the given range without error.

Link: http://lkml.kernel.org/r/1519926220-7453-4-git-send-email-longman@redhat.com
Signed-off-by: Waiman Long <longman@redhat.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Cc: Kees Cook <keescook@chromium.org>
Cc: "Luis R. Rodriguez" <mcgrof@kernel.org>
Cc: Manfred Spraul <manfred@colorfullife.com>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>